Created Answer class with data members answerID, score, isAccepted, body, and authorDisplayName.
Modified Question class by adding new data member, an ArrayList of Answer objects. The list stores all the answers to the specific question.

Created MyAnswerAdapter class.

ViewQuestionActivity now displays all the answers to a question in a ListView manner. Author, votes, and body of the answer are displayed for now.

Unfortunately, the Markdown formatting used by Stack Overflow has not yet been successfully parsed, leading to very unreadable output in the body field of the questions and their answers.