Added caching feature
The app now caches the last 10 queries sent by the user. The json string corresponding to these queries are stored in files inside the application's cache directory. The CacheHandler helper class is used to maintain the cache directory and the files inside it.

For details on how the cache is maintained, see the comment at the top of the CacheHandler class code.

The cache has not been tested for all cases, but for now it retrieves data from cache files when internet connection is not available or the server responds with a null string.